requirements.txt
pandas
openpyxl
python-dotenv
requests
openai
langchain
langchain-openai
langchain-community
langchain-core
flask

